src/config-loader/requirements.txt
pyyaml>=6.0.0
httpx>=0.26.0
